+++ package.json
@@ -15,7 +15,7 @@
     "build": "NODE_ENV=production npm run webpack",
     "clean": "rimraf $npm_package_config_buildDir && mkdir $npm_package_config_buildDir",
     "env": "env",
-    "lint": "eslint --ext .js --ext .jsx ./app ./webpack && echo No linting errors.",
+    "lint": "standard",
     "prebuild": "npm run clean",
     "pretest": "npm install && npm run lint",
     "pretest-travis": "npm install && npm run lint",
@@ -112,6 +112,7 @@
     "sass-loader": "^3.2.0",
     "sinon": "^1.15.4",
     "source-map-support": "^0.4.0",
+    "standard": "^8.0.0",
     "style-loader": "^0.13.1",
     "template-html-loader": "0.0.3",
     "url-loader": "^0.5.7",
